LayoutTests/imported/w3c/web-platform-tests/infrastructure/server/http2-websocket.sub.h2.any-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'wss://localhost:9000/echo' failed: Could not connect to the server.

FAIL WSS over h2 assert_true: expected true got false
